+++ public/locales/en_US/optimizerTab.yaml
@@ -776,6 +776,8 @@ ExpandedDataPanel:
       ELATION: ELATION
     ActionLabel: COMBO ACTION
     FilterLabel: FILTER BY ABILITY
+    DefaultAction: Default
+    SummaryLabel: BUFF TOTALS
     Stats:
       CompositeLabels:
         Label: '{{prefix}} {{suffix}}'
