+++ src/lib/characterPreview/buffsAnalysis/StatSummary.tsx
@@ -1,0 +1,195 @@
+import { Flex, theme } from 'antd'
+import { DAMAGE_TAG_ENTRIES, TagColorEntry } from 'lib/characterPreview/buffsAnalysis/abilityColors'
+import { CardHeader } from 'lib/characterPreview/buffsAnalysis/BuffGroup'
+import { DesignContext, ellipsisStyle, getCardStyle, getIconStyle } from 'lib/characterPreview/buffsAnalysis/designContext'
+import { buffMatchesFilter } from 'lib/characterPreview/buffsAnalysis/FilterBar'
+import { getStatConfig, renderPill, translatedLabel } from 'lib/characterPreview/buffsAnalysis/BuffRow'
+import { BUFF_TYPE } from 'lib/optimization/buffSource'
+import { Buff } from 'lib/optimization/basicStatsArray'
+import { newStatsConfig } from 'lib/optimization/engine/config/statsConfig'
+import { DamageTag } from 'lib/optimization/engine/config/tag'
+import { BuffGroups } from 'lib/simulations/combatBuffsAnalysis'
+import { currentLocale } from 'lib/utils/i18nUtils'
+import { TsUtils } from 'lib/utils/TsUtils'
+import React, { useContext } from 'react'
+import { useTranslation } from 'react-i18next'
+
+const GROUP_ORDER: BUFF_TYPE[] = [
+  BUFF_TYPE.PRIMARY,
+  BUFF_TYPE.SETS,
+  BUFF_TYPE.CHARACTER,
+  BUFF_TYPE.LIGHTCONE,
+]
+
+const STAT_ORDER = new Map(Object.keys(newStatsConfig).map((key, i) => [key, i]))
+
+type StatSumContribution = {
+  value: number
+  damageTags?: number
+}
+
+type StatSum = {
+  stat: string
+  label: string
+  total: number
+  count: number
+  percent: boolean
+  contributions: StatSumContribution[]
+}
+
+export function collectAllBuffs(buffGroups: BuffGroups): Buff[] {
+  const all: Buff[] = []
+  for (const buffType of GROUP_ORDER) {
+    const groupMap = buffGroups[buffType]
+    if (!groupMap) continue
+    for (const buffs of Object.values(groupMap)) {
+      all.push(...buffs)
+    }
+  }
+  return all
+}
+
+export { GROUP_ORDER }
+
+function getBuffStatKey(buff: Buff): string {
+  return buff.memo ? `memo:${buff.stat}` : buff.stat
+}
+
+export function computeStatSums(buffs: Buff[], filter: DamageTag | null): StatSum[] {
+  // First pass: collect all stat keys in insertion order (unfiltered) for stable ordering
+  const keyOrder: string[] = []
+  const keySet = new Set<string>()
+  for (const buff of buffs) {
+    const config = getStatConfig(buff.stat)
+    if (!config || config.bool) continue
+    const key = getBuffStatKey(buff)
+    if (!keySet.has(key)) {
+      keySet.add(key)
+      keyOrder.push(key)
+    }
+  }
+
+  // Second pass: compute sums with filter applied
+  const sumMap = new Map<string, StatSum>()
+  for (const buff of buffs) {
+    if (!buffMatchesFilter(buff, filter)) continue
+    const config = getStatConfig(buff.stat)
+    if (!config || config.bool) continue
+
+    const key = getBuffStatKey(buff)
+    const contribution: StatSumContribution = {
+      value: buff.value,
+      damageTags: buff.damageTags,
+    }
+
+    const existing = sumMap.get(key)
+    if (existing) {
+      existing.total += buff.value
+      existing.count++
+      existing.contributions.push(contribution)
+    } else {
+      sumMap.set(key, {
+        stat: buff.stat,
+        label: translatedLabel(buff.stat, buff.memo),
+        total: buff.value,
+        count: 1,
+        percent: !config.flat,
+        contributions: [contribution],
+      })
+    }
+  }
+
+  // Return in stable key order, filtering out zero-total entries
+  return keyOrder
+    .map((key) => sumMap.get(key))
+    .filter((s): s is StatSum => s != null && s.total !== 0)
+    .sort((a, b) => (STAT_ORDER.get(a.stat) ?? 999) - (STAT_ORDER.get(b.stat) ?? 999))
+}
+
+function getContributionTagPills(contributions: StatSumContribution[]): TagColorEntry[] {
+  const hasAll = contributions.some((c) => c.damageTags == null)
+  const specificTags = new Set<DamageTag>()
+  for (const c of contributions) {
+    if (c.damageTags != null) {
+      for (const entry of DAMAGE_TAG_ENTRIES) {
+        if (c.damageTags & entry.tag) specificTags.add(entry.tag)
+      }
+    }
+  }
+
+  const pills: TagColorEntry[] = []
+  if (hasAll) pills.push({ color: '#8c8c8c', key: 'ALL' })
+  for (const entry of DAMAGE_TAG_ENTRIES) {
+    if (specificTags.has(entry.tag)) pills.push({ color: entry.color, key: entry.key })
+  }
+  return pills
+}
+
+function formatStatSum(sum: StatSum): string {
+  if (sum.percent) {
+    return TsUtils.precisionRound(sum.total * 100, 2).toLocaleString(currentLocale()) + ' %'
+  }
+  return TsUtils.precisionRound(sum.total, 0).toLocaleString(currentLocale())
+}
+
+function SummaryTagPills(props: { contributions: StatSumContribution[] }) {
+  const { t } = useTranslation('optimizerTab', { keyPrefix: 'ExpandedDataPanel.BuffsAnalysisDisplay.DamageTags' })
+  const pills = getContributionTagPills(props.contributions)
+  if (pills.length === 0) return null
+
+  return (
+    <Flex gap={2} style={{ flexShrink: 0 }}>
+      {pills.map((p) => renderPill(p.key, p.color, t(p.key)))}
+    </Flex>
+  )
+}
+
+export function StatSummaryTable(props: { sums: StatSum[]; avatarSrc: string }) {
+  const options = useContext(DesignContext)
+  const { token } = theme.useToken()
+  const { t } = useTranslation('optimizerTab', { keyPrefix: 'ExpandedDataPanel.BuffsAnalysisDisplay' })
+  const cardStyleProp = getCardStyle(options, token)
+  const iconStyle = getIconStyle(options)
+
+  return (
+    <Flex align='center' gap={0} style={cardStyleProp}>
+      <img src={props.avatarSrc} style={iconStyle} />
+      <Flex vertical gap={0} style={{ flex: 1, overflow: 'hidden' }}>
+        <CardHeader label={t('SummaryLabel')} />
+        {props.sums.map((sum, i) => (
+          <Flex
+            key={sum.label}
+            align='center'
+            gap={6}
+            style={{
+              padding: `0 ${options.rowPaddingX}px`,
+              height: options.rowHeight,
+              lineHeight: `${options.rowHeight}px`,
+              borderBottom: i < props.sums.length - 1 ? `1px solid ${options.borderColor}` : undefined,
+            }}
+          >
+            <span style={{ minWidth: 60, fontSize: options.fontSize, textWrap: 'nowrap' }}>
+              {formatStatSum(sum)}
+            </span>
+
+            <span style={{ minWidth: 150, ...ellipsisStyle(options.fontSize) }}>
+              {'∑ ' + sum.label}
+            </span>
+
+            <SummaryTagPills contributions={sum.contributions} />
+
+            <span style={{
+              marginLeft: 'auto',
+              color: `rgba(255,255,255,${options.sourceOpacity / 100})`,
+              fontSize: options.fontSize,
+              textWrap: 'nowrap',
+              flexShrink: 0,
+            }}>
+              {'x' + sum.count}
+            </span>
+          </Flex>
+        ))}
+      </Flex>
+    </Flex>
+  )
+}
