+++ src/lib/characterPreview/buffsAnalysis/designContext.ts
@@ -1,0 +1,49 @@
+import React, { createContext } from 'react'
+import { DamageTag } from 'lib/optimization/engine/config/tag'
+import { cardShadow } from 'lib/tabs/tabOptimizer/optimizerForm/layout/FormCard'
+
+export type DesignOptions = {
+  rowHeight: number
+  iconSize: number
+  sourceOpacity: number
+  tintIntensity: number
+  panelWidth: number
+  rowPaddingX: number
+  fontSize: number
+  borderColor: string
+  cardPadding: number
+}
+
+export const BORDER_RADIUS = 5
+export const GROUP_SPACING = 10
+export const PILL_SIZE = { padding: '0 4px', fontSize: 9, lineHeight: '16px' } as const
+
+export const DEFAULT_OPTIONS: DesignOptions = {
+  rowHeight: 26,
+  iconSize: 48,
+  sourceOpacity: 45,
+  tintIntensity: 15,
+  panelWidth: 600,
+  rowPaddingX: 6,
+  fontSize: 12,
+  borderColor: '#ffffff0f',
+  cardPadding: 3,
+}
+
+export const DesignContext = createContext<DesignOptions>(DEFAULT_OPTIONS)
+export const FilterContext = createContext<DamageTag | null>(null)
+
+export const ellipsisStyle = (fontSize: number): React.CSSProperties => ({
+  overflow: 'hidden',
+  textOverflow: 'ellipsis',
+  textWrap: 'nowrap',
+  fontSize,
+})
+
+export function getCardStyle(options: DesignOptions, token: { colorBgContainer: string }): React.CSSProperties {
+  return { borderRadius: BORDER_RADIUS, overflow: 'hidden', padding: options.cardPadding, backgroundColor: token.colorBgContainer, boxShadow: cardShadow }
+}
+
+export function getIconStyle(options: DesignOptions): React.CSSProperties {
+  return { width: options.iconSize, height: options.iconSize, flexShrink: 0, objectFit: 'cover', margin: '0 4px' }
+}
